donut
[Verse 1]
All day and all night
We lay about and wonder what we're gonna do
We speak to ourselves
Havin' a conversation and wonderin' who we're talking to
Then sadness takes over
And sadness wonders what we're gonna do
And sadness comes closer
Then sadness comes to sing a song for you

[Verse 2]
Oh, what a wonderful world (Oh-oh-oh)
Oh, how you've got to have more
When darkness covers the Earth
And I know it's the end for sure

[Chorus]
Here comes the sun
Here comes the sun
Here comes the sun
Bringing you love and shining on everyone
Here comes the sun
Here come the raging sun
And it's raging on and so we rage as one
On until the day is done

[Verse 3]
High on the mountain
The postman comes and says there's still no news
Deep in the city, oh yeah
The buildings wonder if they'll ever be built

[Verse 4]
Oh, what an interesting day
Looks like there's nobody in my way
But the road's gonna end
And the storm's comin' in
Looks like nobody wins today

[Chorus]
Here comes the sun
Here comes the sun
Here comes the sun
Bringing you love and shining on everyone
Here comes the sun
Here come the blazing amazing miracle sun
And it's raging on and still, we'll blaze along
On until the day is done

[Bridge]
'Til the day is won and done
Gonna rage on into the sun
Call up your friends and neighbours
Tell 'em a new day has just begun
I hope you know that there's no stranger strange enough
Tell all your friends and neighbours
You're no stranger to fallin' in love
I'm just fallin' in love
Stop, time's up
Is it really over?
Time waits for no one
(And it doesn't wait now either)
And it can't explain the way that I'm feelin'
(Ding, ding, ding, dang, ding, ding)
Put it into words and then
Send it all over the world!

[Chorus]
Here comes the sun
Here comes the sun
Here comes the sun
Bringing you love and shining on everyone
Here comes the sun
Here comes the raging sun, yeah
And it's raging on and so we rage as one
On until the day is done
Oh-oh, until the day is done

[Outro]
If this song ain't good enough I'll erase it